use-tipoPropiedad
